Fix issue launch command line
- The command line wasn't being properly build with nested referenced projects. Because the loaded assemblies path list was incomplete some assemblyes were missing causing the debugger engine to crash VS
- Create new class ReferenceCrawler. Moved relevant methods there to allow reuse from deploy and launch providers
