+++ source/VisualStudio.Extension/VisualStudio.Extension.csproj
@@ -144,6 +144,7 @@
     <Compile Include="ToolWindow.DeviceExplorer\ViewModel\DeviceExplorerViewModel.cs" />
     <Compile Include="ToolWindow.DeviceExplorer\ViewModel\IDataService.cs" />
     <Compile Include="ToolWindow.DeviceExplorer\ViewModel\ViewModelLocator.cs" />
+    <Compile Include="Utilities\ReferenceCrawler.cs" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="..\packages\nanoFramework.CoreLibrary.1.0.0-preview062\nanoFramework.CoreLibrary.1.0.0-preview062.nupkg">
